pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rdo-agent"
version = "0.1.0"
description = "Agente Forense de RDO — Sistema multi-agente para geração de Relatórios Diários de Obra"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Lucas Ferreira Leite", email = "lucas@valenobre.com.br" }
]
license = { text = "Proprietary" }

dependencies = [
    # Core
    "python-dateutil>=2.8.2",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",

    # Metadata extraction
    "pymediainfo>=6.1.0",

    # Task scheduling
    "apscheduler>=3.10.0",

    # APIs (para sprints 2+)
    "openai>=1.12.0",
    "anthropic>=0.18.0",

    # Opentimestamps
    "opentimestamps-client>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "ipython>=8.18.0",
]

[project.scripts]
rdo-agent = "rdo_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
